branches/main: merge from 64 bits branch
git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/main@821 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ libs/include/cisstMultiTask/mtsStateIndex.h
@@ -54,7 +54,7 @@ class CISST_EXPORT mtsStateIndex : public mtsGenericObject {
 
  public:
     /*! TimeTicks are typedef'ed as unsigned long */
-    typedef unsigned long TimeTicksType;
+    typedef unsigned long long int TimeTicksType;
     
  private:
 	/*! The index into the set of circular buffers corresponding to
